BUILD: atomic: make the old HA_ATOMIC_LOAD() support const pointers
We have an implementation of atomic ops for older versions of gcc that
do not provide the __builtin_* API (< 4.4). Recent changes to the pools
broke that in pool_releasable() by having a load from a const pointer,
which doesn't work there due to a temporary local variable that is
declared then assigned. Let's make use of a compount statement to assign
it a value when declaring it.

There's no need to backport this.

(cherry picked from commit 8da23393a1ec3f24dcd408f4b35fdd3d9a141a5f)
[wt: 2.5 on gcc-4.4 is reportedly affected by this as well, e.g. in
 freq_ctr_total()]
Signed-off-by: Willy Tarreau <w@1wt.eu>