BUG/MEDIUM: resolvers: Really ignore trailing dot in domain names
When a string is converted to a domain name label, the trailing dot must be
ignored. In resolv_str_to_dn_label(), there is a test to do so. However, the
trailing dot is not really ignored. The character itself is not copied but
the string index is still moved to the next char. Thus, this trailing dot is
counted in the length of the last encoded part of the domain name. Worst,
because the copy is skipped, a garbage character is included in the domain
name.

This patch should fix the issue #1528. It must be backported as far as 2.0.

(cherry picked from commit 0a82cf4c1662b8ab00036f65b5e3543a9c1a6ff5)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>
(cherry picked from commit 11fc8f8632d6bf6b71d8dcc505b3fab0d1abf714)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 5c3e686d11b3232514006877f63333cdd9b08ad0)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>