CLEANUP: opentracing: added flt_ot_smp_init() function
The flt_ot_smp_init() function has been added to make initializing the
sample structure easier.  The contents of the structure in question are
set in several places in the source of the OpenTracing filter, so it is
better to do this in one place.

This patch must be backported as far as 2.4.

(cherry picked from commit 854efefd6cf43018efcdeb0da2c06cc19fd00980)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>