BUG/MEDIUM: httpclient/lua: infinite appctx loop with POST
When POSTing a request with a payload, and reusing the same httpclient
lua instance, one could encounter a spinning of the httpclient appctx.

Indeed the sent counter is not reset between 2 POSTs and the condition
for sending the EOM flag is never met.

Must fixed issue #1593.

To be backported in 2.5.

(cherry picked from commit 10a37360c8c5dfc4a65aba81f67ae63fe5fa2161)
[wla: context changes]
Signed-off-by: William Lallemand <wlallemand@haproxy.org>