BUG/MINOR: hlua: Set conn-stream/channel EOI flags at the end of request
This bug is the same than for the HTTP client. See "BUG/MINOR: httpclient:
Set conn-stream/channel EOI flags at the end of request" for details.

This patch must be backported as far as 2.0. But only CF_EOI must be set
because applets are not attached to a conn-stream on older versions.

(cherry picked from commit d8d2708cfef417e3c4e2f627089b49e030ebbe80)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>