BUG/MINOR: cache: Set conn-stream/channel EOI flags at the end of request
This bug is the same than for the HTTP client. See "BUG/MINOR: httpclient:
Set conn-stream/channel EOI flags at the end of request" for details.

Note that because a filter is always attached to the stream when the cache
is used, there is no issue because there is no direct forwarding in this
case. Thus the stream analyzers are able to see the HTX_FL_EOM flag on the
HTX messge.

This patch must be backported as far as 2.0. But only CF_EOI must be set
because applets are not attached to a conn-stream on older versions.

(cherry picked from commit dbf1e88e87748aef2db25fcdfeb22a4005908d41)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 3fcefd16e3f9493bd56c00e134c5831fbfd08000)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit ba4761221bdb00c7057ea6919bae9f938bfaba18)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>